feat: Add address and bank info to Agency type
This commit introduces the `Address` and `BankInfo` interfaces and integrates them into the `Agency` and `User` types. This change supports storing detailed contact and financial information for agencies, as well as basic address details for clients.

The `AuthContext` and `DataContext` have been updated to handle the mapping and persistence of this new data. Additionally, new features for updating passwords, deleting accounts, and uploading images have been added to `AuthContext` to support broader user management capabilities.

UI components in `AgencyDashboard` and `ClientDashboard` will need to be updated to reflect these changes and utilize the new fields.